Move liquidity mining provider
